+++ src/MagicTrick.java
@@ -24,7 +24,7 @@ public class MagicTrick
 		{
 			fileStream = new FileInputStream("A-small.in");
 			fileReader = new BufferedReader(new InputStreamReader(fileStream, Charset.forName("UTF-8")));
-			numEntries = Byte.parseByte(fileReader.readLine());
+			numEntries = Integer.parseInt(fileReader.readLine());
 			
 			for (int i = 1; i <= numEntries; i++)
 			{
